some improvement

+++ upload.php
@@ -8,9 +8,13 @@ if(isset($_FILES)){
 		header("Location: index.php?err=1");		
 		return;
 	}
+	//get song name from filename
 	$song_filename = str_replace(",","_",$song_filename);
-	$dest = $song_folder.$song_filename;
 	$song_name = str_replace($ext,"",$song_filename);
+	//get md5
+	$md5_src = md5_file($src);
+	$song_filename = $md5_src.$ext;
+	$dest = $song_folder.$song_filename;
 	if(isset($_POST["newsong"])){
 		if(!empty($_POST["newsong"])){
 			$song_name = $_POST["newsong"];
@@ -20,15 +24,18 @@ if(isset($_FILES)){
 		mkdir($song_folder);
 		chmod($song_folder, $dir_mode);
 	}
-	if (move_uploaded_file($src, $dest)) { 
-		chmod($dest, $file_mode);
-		$fp = fopen($list_file,"a");
-		flock($fp, LOCK_SH);
-		fwrite($fp, $song_name.",".$song_filename."\n");
-		flock($fp, LOCK_UN);
-		fclose($fp);
-		@unlink($_FILES["songfile"]["tmp_name"]); 
-	} 
+	$file_exists=0;
+	if(!file_exists($dest)){
+		if (move_uploaded_file($src, $dest)) { 
+			chmod($dest, $file_mode);
+			@unlink($_FILES["songfile"]["tmp_name"]); 
+		}
+	}
+	$fp = fopen($list_file,"a");
+	flock($fp, LOCK_SH);
+	fwrite($fp, $song_name.",".$song_filename."\n");
+	flock($fp, LOCK_UN);
+	fclose($fp);
 }
-header("Location: /");
+header("Location: index.php");
 ?>
